gdata: GetAvailableSpace should return int64
We should use 64-bit integer for the arguments bytes_total and bytes_used in
the GetAvailableSpace callback.

Updating unit tests with >4GB sizes to verify that the sizes are not clamped.

BUG=chromium-os:30011
TEST=Verify that gdata file system unit tests pass.


Review URL: http://codereview.chromium.org/10174033

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@134060 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 81663c15c7771132c7344ffd11fdff82fac3c603

+++ chrome/browser/chromeos/gdata/gdata_file_system.h
@@ -79,8 +79,8 @@ typedef base::Callback<void(base::PlatformFileError error,
 
 // Used to get available space for the account from GData.
 typedef base::Callback<void(base::PlatformFileError error,
-                            int bytes_total,
-                            int bytes_used)>
+                            int64 bytes_total,
+                            int64 bytes_used)>
     GetAvailableSpaceCallback;
 
 // Callback type for DocumentServiceInterface::ResumeUpload.
